Add support for plugins in chatcli
Adds the `--plugin` option to the `add` command that activates plugins.
Currently, there is only one plugin, `pyeval`, that allows evaluating
Python expressions by prefixing the input with `evaluate: `. The
`answer` function also receives a list of activated plugins, and it uses
the `pyeval` plugin to evaluate code blocks.

Tests are also updated to test this new feature.